# incident-dashboard/backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "异常事件登记台 API",
    "description": "事件管理 REST API，支持事件 CRUD、状态流转、乐观锁、幂等控制、超时提醒、数据导出等功能。",
    "version": "1.0.0",
    "contact": {
      "name": "XiaopengHSE",
      "url": "https://github.com/XiaopengHSE/incident-management"
    }
  },
  "servers": [
    {
      "url": "https://backend-c8ax.vercel.app",
      "description": "Vercel 生产环境"
    },
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    }
  ],
...
  "components": {
    "schemas": {
      "Incident": {
        "type": "object",
        "description": "事件实体",
        "properties": {
          "id": {
            "type": "string",
            "description": "事件唯一ID",
            "example": "inc_mtb9eyyu_uwlp0a"
          },
          "title": {
            "type": "string",
            "description": "事件标题",
            "maxLength": 200,
            "example": "支付模块异常"
          },
          "severity": {
            "type": "string",
            "description": "严重度",
            "enum": ["紧急", "高", "中", "低"],
            "example": "高"
          },
          "assignee": {
            "type": "string",
            "description": "负责人",
            "maxLength": 50,
            "example": "张三"
          },
          "status": {
            "type": "string",
            "description": "处理状态",
            "enum": ["待处理", "处理中", "已解决", "已关闭"],
            "example": "待处理"
          },
          "remark": {
            "type": "string",
            "description": "备注",
            "maxLength": 1000,
            "example": "第三方支付网关超时导致订单状态不一致"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间",
            "example": "2026-08-27T08:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "最后更新时间",
            "example": "2026-08-27T08:00:00.000Z"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "description": "关闭时间（未关闭为 null）",
            "example": null
          },
          "history": {
            "type": "array",
            "description": "状态变更历史",
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            }
          },
          "idempotencyKey": {
            "type": "string",
            "description": "幂等键（防重复提交）",
            "example": "create_支付模块异常_张三"
          },
          "version": {
            "type": "integer",
            "description": "乐观锁版本号",
            "example": 1
          }
        },
        "required": ["id", "title", "severity", "status", "createdAt", "updatedAt", "version"]
      },
      "HistoryEntry": {
        "type": "object",
        "description": "状态变更记录",
        "properties": {
          "status": {
            "type": "string",
            "description": "变更后的状态",
            "example": "处理中"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "description": "变更时间",
            "example": "2026-08-27T08:00:00.000Z"
          },
          "remark": {
            "type": "string",
            "description": "变更说明",
            "example": "状态变更: 待处理 → 处理中"
          }
        }
      },
      "CreateIncidentRequest": {
        "type": "object",
        "description": "创建事件请求体",
        "properties": {
          "title": {
            "type": "string",
            "description": "事件标题（必填，1-200字）",
            "minLength": 1,
            "maxLength": 200,
            "example": "支付模块异常"
          },
          "severity": {
            "type": "string",
            "description": "严重度（必填）",
            "enum": ["紧急", "高", "中", "低"],
            "example": "高"
          },
          "assignee": {
            "type": "string",
            "description": "负责人（可选，最长50字）",
            "maxLength": 50,
            "example": "张三"
          },
          "status": {
            "type": "string",
            "description": "初始状态（必填）",
            "enum": ["待处理", "处理中", "已解决", "已关闭"],
            "example": "待处理"
          },
          "remark": {
            "type": "string",
            "description": "备注（可选，最长1000字）",
            "maxLength": 1000,
            "example": "第三方支付网关超时导致订单状态不一致"
          },
          "idempotencyKey": {
            "type": "string",
            "description": "幂等键（可选，防重复提交。同一 key 重复提交会返回已创建的事件）",
            "example": "create_支付模块异常_张三"
          }
        },
        "required": ["title", "severity", "status"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "错误响应",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "错误码",
                "enum": ["VALIDATION", "FORBIDDEN", "CONFLICT", "IDEMPOTENT", "NOT_FOUND", "INTERNAL"],
                "example": "VALIDATION"
              },
              "message": {
                "type": "string",
                "description": "错误描述",
                "example": "事件标题不能为空; 处理状态必须是: 待处理/处理中/已解决/已关闭"
              },
              "reqId": {
                "type": "string",
                "description": "请求ID（用于追踪）",
                "example": "req_mtb9eyyu_uwlp"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/incidents": {
      "get": {
        "tags": ["事件管理"],
        "summary": "查询事件列表",
        "description": "支持按状态、严重度、负责人、关键词筛选，默认按创建时间倒序排列。",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "按处理状态筛选",
            "schema": {
              "type": "string",
              "enum": ["待处理", "处理中", "已解决", "已关闭"]
            }
          },
          {
            "name": "severity",
            "in": "query",
            "description": "按严重度筛选",
            "schema": {
              "type": "string",
              "enum": ["紧急", "高", "中", "低"]
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "description": "按负责人筛选（模糊匹配）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "关键词搜索（匹配标题、备注、负责人）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excludeClosed",
            "in": "query",
            "description": "排除已关闭事件",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "incidents": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Incident" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["事件管理"],
        "summary": "创建事件",
        "description": "创建新的事件。支持幂等控制：传入 idempotencyKey 时，重复提交不会创建新事件，而是返回已有事件。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateIncidentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "incident": { "$ref": "#/components/schemas/Incident" }
                  }
                }
              }
            }
          },
          "200": {
            "description": "幂等重复请求（返回已创建的事件）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "incident": { "$ref": "#/components/schemas/Incident" },
                    "_idempotent": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数校验失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}